actix_security_core::http::security

Trait Authenticator

pub trait Authenticator {
    // Required method
    fn get_user(&self, req: &ServiceRequest) -> Option<User>;
}

Trait for extracting user identity from an HTTP request.

Spring Equivalent

AuthenticationProvider / UserDetailsService

Implementation Note

Returns an owned User so it can be stored in request extensions for access by handlers.

Required Methods

fn get_user(&self, req: &ServiceRequest) -> Option<User>

Attempts to authenticate the request and returns the user if successful.